requirements.txt
pytelegrambotapi
apscheduler
simplematrixbotlib
requests~=2.30.0
python-dotenv